requirements.txt
absl-py==2.1.0
asttokens==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
executing==2.1.0
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
h5py==3.12.1
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.8.0
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.1
opencv-python==4.10.0.84
optree==0.13.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
pyzmq==26.2.0
rich==13.9.4
scikit-learn==1.6.0
scipy==1.15.0
setuptools==75.7.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.6.0+cu124
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
